requirements.txt
moviepy>=1.0.3,<2.0
ffmpeg-python>=0.2.0
Pillow>=10.0
numpy>=1.24
google-generativeai>=0.5
edge-tts>=6.1
requests>=2.31
aiohttp>=3.9
google-auth>=2.22
google-auth-oauthlib>=1.0
google-api-python-client>=2.90
pysrt>=1.1.2
pyyaml>=6.0
tqdm>=4.65
python-dateutil>=2.8
